deprecate asdf.resolver submodule
