+++ CHANGES.rst
@@ -20,6 +20,7 @@ The ASDF Standard is at v1.6.0
 - Add AsdfDeprecationWarning to AsdfInFits. Support for reading and
   writing ASDF in fits files is being moved to `stdatamodels
   <https://github.com/spacetelescope/stdatamodels>`_. [#1337]
+- Add AsdfDeprecationWarning to asdf.resolver [#1362]
 
 2.14.3 (2022-12-15)
 -------------------
