+++ asdf/extension/_legacy.py
@@ -2,7 +2,8 @@ import abc
 import warnings
 from functools import lru_cache
 
-from asdf import resolver, types
+from asdf import _resolver as resolver
+from asdf import types
 from asdf.exceptions import AsdfDeprecationWarning
 from asdf.type_index import AsdfTypeIndex
