+++ asdf/tests/helpers.py
@@ -23,12 +23,12 @@ import yaml
 
 import asdf
 from asdf import generic_io, versioning
+from asdf._resolver import Resolver, ResolverChain
 from asdf.asdf import AsdfFile, get_asdf_library_info
 from asdf.block import Block
 from asdf.constants import YAML_TAG_PREFIX
 from asdf.exceptions import AsdfConversionWarning
 from asdf.extension import default_extensions
-from asdf.resolver import Resolver, ResolverChain
 from asdf.tags.core import AsdfObject
 from asdf.versioning import (
     AsdfVersion,
