+++ asdf/tests/test_deprecated.py
@@ -18,3 +18,12 @@ def test_asdf_in_fits_import_warning():
         del sys.modules["asdf.fits_embed"]
     with pytest.warns(AsdfDeprecationWarning, match="AsdfInFits has been deprecated.*"):
         import asdf.fits_embed  # noqa: F401
+
+
+def test_resolver_module_deprecation():
+    with pytest.warns(AsdfDeprecationWarning, match="^asdf.resolver is deprecated.*$"):
+        # importlib.reload doesn't appear to work here likely because of the
+        # sys.module and __file__ changes in asdf.resolver
+        if "asdf.resolver" in sys.modules:
+            del sys.modules["asdf.resolver"]
+        import asdf.resolver  # noqa: F401
