+++ asdf/tests/test_schema.py
@@ -7,7 +7,8 @@ from jsonschema import ValidationError
 from numpy.testing import assert_array_equal
 
 import asdf
-from asdf import config_context, constants, extension, get_config, resolver, schema, tagged, types, util, yamlutil
+from asdf import _resolver as resolver
+from asdf import config_context, constants, extension, get_config, schema, tagged, types, util, yamlutil
 from asdf.exceptions import AsdfConversionWarning, AsdfDeprecationWarning, AsdfWarning
 from asdf.tests import helpers
 from asdf.tests.objects import CustomExtension
